requirements.txt
pandas
yfinance
schedule
PyYAML
python-dotenv # Optional, but good practice for API keys later
pyarrow  # For Parquet
fredapi  # For FRED data